Add new features, code clean up
Converted tabs into spaces
Converted comparisons to None
Added created_files property which records the files that have been written
Added clean_files in constructor which defaults to false
Added .clean() which will remove the created files if clean_files is true
Added __enter__ and __exit__ so that program can be called using with and automatically clean files on exit
Added id type(imgfilepath) is numpy.ndarray: so that sew can be called with a numpy.ndarray data and not just the location of a fits file. The image data is written out in workdir with the name imgname if set or sew_image.